requirements.txt
Django==5.2.12
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.29.0
celery==5.6.3
redis==7.4.0
python-dotenv==1.1.0
stripe==15.0.1
requests==2.32.5
pytest==9.0.2
pytest-django==4.12.0
